finally fixed projectile damage error

+++ Assets/Scripts/Gameplay/General/Status/DamageComponent.cs
@@ -3,6 +3,6 @@ using Unity.Entities;
 public struct DamageComponent : IComponentData
 {
     public uint damage;
-    public byte pierce;
+    public ushort pierce;
     //public DamageType damageType;
 }
